+++ src/main/java/chess/board/piece/Bishop.java
@@ -1,8 +1,8 @@
-package chess.piece;
+package chess.board.piece;
 
 import chess.board.Variation;
 
-public class Rook implements Piece {
+public class Bishop implements Piece {
     @Override
     public boolean canMove(final Variation variation) {
         return false;
