+++ src/main/java/chess/board/piece/Queen.java
@@ -1,8 +1,8 @@
-package chess.piece;
+package chess.board.piece;
 
 import chess.board.Variation;
 
-public class Bishop implements Piece {
+public class Queen implements Piece {
     @Override
     public boolean canMove(final Variation variation) {
         return false;
