+++ src/main/java/chess/board/piece/Team.java
@@ -1,0 +1,4 @@
+package chess.board.piece;
+
+public enum Team {
+}
