--- src/main/java/chess/piece/Team.java
@@ -1,4 +1,0 @@
-package chess.piece;
-
-public enum Team {
-}
